Fix circular dependencies in src/acl
